Refer to .json file, not .xml, and fix serialisation
