+++ apps/common-lib/src/main/scala/com/gu/typerighter/rules/BucketRuleResource.scala
@@ -3,7 +3,7 @@ package com.gu.typerighter.rules
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.{ObjectMetadata, PutObjectRequest}
 import com.gu.typerighter.lib.SafeXMLParser
-import com.gu.typerighter.model.{CheckerRuleResource, WordlistsToNotPublish, WordTag}
+import com.gu.typerighter.model.{CheckerRuleResource, TaggedWordlist, WordTag}
 import play.api.Logging
 import play.api.data.FormError
 import play.api.libs.json.Json
@@ -14,7 +14,7 @@ import scala.util.{Failure, Success, Try}
 class BucketRuleResource(s3: AmazonS3, bucketName: String, stage: String) extends Logging {
   private val RULES_KEY = s"$stage/rules/typerighter-rules.json"
   private val DICTIONARY_KEY = s"$stage/dictionary/typerighter-dictionary.xml"
-  private val WORDS_TO_NOT_PUBLISH_KEY = s"$stage/dictionary/words-to-not-publish.xml"
+  private val WORDS_TO_NOT_PUBLISH_KEY = s"$stage/dictionary/words-to-not-publish.json"
 
   def putRules(ruleResource: CheckerRuleResource): Either[Exception, Unit] = {
     val ruleJson = Json.toJson(ruleResource)
@@ -66,12 +66,12 @@ class BucketRuleResource(s3: AmazonS3, bucketName: String, stage: String) extend
       val wordsStream = words.getObjectContent()
       val wordlistsToNotPublishJson = Json.parse(wordsStream)
       words.close()
-      wordlistsToNotPublishJson.as[WordlistsToNotPublish]
+      wordlistsToNotPublishJson.as[List[TaggedWordlist]]
     })
 
     wordlistsToNotPublish match {
       case Success(wordsToNotPublish) =>
-        wordsToNotPublish.rules.flatMap(wordlist =>
+        wordsToNotPublish.flatMap(wordlist =>
           wordlist.words.map(word => WordTag(word, wordlist.tag))
         )
       case Failure(_) =>
